Move the information of the doors in the walls to another array.
The doors have different properties than the other objects
attached to the walls. The doors are present at the two sides of the wall,
and the position is fixed. I have moved the doors to a 'doors' array,
and the 'objects' array will be loaded with the other objects.
I think the code will be simpler in this way.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>