Start movement of character in play mode
Now the character does not walk all time left and right,
but wait to a click, and go to the place,
if the click is over a lateral wall or over a door, the character walk and later
the map change to the new selected vision.

Is not working all right now. When the character go to the left
the svg is not rendered. A posible solution using to ImageSurfaces
to cache te rendering is commented in character.py

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>